Rename OutputPayload to Output::Payload
