+++ lib/tapping_device/trackers/mutation_tracker.rb
@@ -58,12 +58,12 @@ class TappingDevice
 
         additional_keys = @latest_instance_variables.keys - @instance_variables_snapshot.keys
         additional_keys.each do |key|
-          changes[key] = {before: OutputPayload::UNDEFINED, after: @latest_instance_variables[key]}
+          changes[key] = {before: Output::Payload::UNDEFINED, after: @latest_instance_variables[key]}
         end
 
         removed_keys = @instance_variables_snapshot.keys - @latest_instance_variables.keys
         removed_keys.each do |key|
-          changes[key] = {before: @instance_variables_snapshot[key], after: OutputPayload::UNDEFINED}
+          changes[key] = {before: @instance_variables_snapshot[key], after: Output::Payload::UNDEFINED}
         end
 
         remained_keys = @latest_instance_variables.keys - additional_keys
